- added nth in dataloader @Robocraft999 - renamed tagreader to id3tagreader - fixed m3u linebreak detection - improved urldecoding
