+++ Cargo.toml
@@ -9,3 +9,4 @@ authors = ["Klassenserver7b <klassenserver7bwin10@gmail.com>", "Konsl <82901383+
 [dependencies]
 id3 = "1.13.1"
 url = "2.5.2"
+percent-encoding = "2.3.1"
