+++ src/dataloading/dataprovider/playlist_dataprovider.rs
@@ -1,5 +1,4 @@
 use crate::model::SongInfo;
 use crate::dataloading::dataprovider::VectorForwardBackward;
 
-
 pub type PlaylistDataProvider<'a> = VectorForwardBackward<'a, SongInfo>;
